# Speed up Enum.empty?/1 x3 times for Map