@josevalim 2020-11-12
We generally don't like to special case our types in Enum because it defeats the point of a protocol if our built-in data types have special powers compared to user ones. We do it only in the cases it is extremely necessary, such as lists.

Although we have been discussing adding new functions to the protocol, especially on the areas we special case today, such as Enum.sum and Enum.to_list. Although I don't think there is something we can do for empty, as most of the cost is on the protocol dispatch. Thank you!