@Strech 2020-11-12
@josevalim Thanks for the explanation. Unfortunately, I didn't know about the discussion around the protocol of the types, so sorry for the useless PR